backend/requirements.txt
Django==3.2
django-environ==0.11.2
django-cors-headers==4.4.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.3.1
requests==2.32.3
